Implementando dao Responsavel e tipo compra

+++ ApiSGCOlimpiada/Models/Acompanhamento.cs
@@ -7,12 +7,12 @@ namespace ApiSGCOlimpiada.Models
 {
     public class Acompanhamento
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Observacao { get; set; }
         public DateTime Date { get; set; }
-        public int StatusId { get; set; }
-        public int UsuarioId { get; set; }
-        public int SolicitacaoCompraId { get; set; }
+        public long StatusId { get; set; }
+        public long UsuarioId { get; set; }
+        public long SolicitacaoCompraId { get; set; }
         public Status Status { get; set; }
         public Usuario Usuario { get; set; }
         public SolicitacaoCompra SolicitacaoCompra { get; set; }
