+++ ApiSGCOlimpiada/Models/Responsavel.cs
@@ -7,10 +7,10 @@ namespace ApiSGCOlimpiada.Models
 {
     public class Responsavel
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string  Nome { get; set; }
         public string  Cargo { get; set; }
-        public int EscolaId { get; set; }
+        public long EscolaId { get; set; }
         public Escola Escola { get; set; }
     }
 }
